changed windows compiler, fixed bug in sv.h, fixed incorrect integer type in file.h, expanded platform layer

+++ src/file_io.h
@@ -7,7 +7,7 @@
 
 // NOTE: interface is coupled with arena. This should potentially be reconsidered in the future
 
-typedef struct file_buffer { void* Buffer; u64 Size; } file_buffer;
+typedef struct file_buffer { void* Buffer; u32 Size; } file_buffer;
 
-bool         WriteEntireFile(char* path, void* buffer, u64 bufferSize);
+bool         WriteEntireFile(char* path, void* buffer, u32 bufferSize);
 file_buffer  ReadEntireFile (char* path, arena* a);
